# packages/shared-proto/openapi.yaml
openapi: 3.0.3
info:
  title: ACE API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /auth/register:
    post:
      deprecated: true
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      deprecated: true
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/me:
    get:
      deprecated: true
      operationId: me
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /auth/refresh:
    post:
      deprecated: true
      operationId: refresh
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Refreshed session (sets auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/logout:
    post:
      deprecated: true
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /auth/logout-all:
    post:
      deprecated: true
      operationId: logoutAll
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out of all sessions (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /student/auth/register:
    post:
      operationId: studentRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /student/auth/login:
    post:
      operationId: studentLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /student/auth/me:
    get:
      operationId: studentMe
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /student/auth/refresh:
    post:
      operationId: studentRefresh
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Refreshed session (sets auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /student/auth/logout:
    post:
      operationId: studentLogout
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /student/auth/logout-all:
    post:
      operationId: studentLogoutAll
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out of all sessions (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /instructor/auth/login:
    post:
      operationId: instructorLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /instructor/auth/me:
    get:
      operationId: instructorMe
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /instructor/auth/refresh:
    post:
      operationId: instructorRefresh
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Refreshed session (sets auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /instructor/auth/logout:
    post:
      operationId: instructorLogout
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /instructor/auth/logout-all:
    post:
      operationId: instructorLogoutAll
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out of all sessions (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /admin/auth/login:
    post:
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /admin/auth/me:
    get:
      operationId: adminMe
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /admin/auth/refresh:
    post:
      operationId: adminRefresh
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Refreshed session (sets auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /admin/auth/logout:
    post:
      operationId: adminLogout
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /admin/auth/logout-all:
    post:
      operationId: adminLogoutAll
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '200':
          description: Logged out of all sessions (clears auth cookies)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: ace_access
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema:
        type: string
      description: Double-submit CSRF token. Must match the ace_csrf cookie.
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          description: Plaintext password (sent over HTTPS in production)
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, instructor, admin]
        createdAt:
          type: string
          description: RFC3339 timestamp
...
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/User'